Fix data synchronize bug

+++ src/sar_localization/src/2D_listener.cpp
@@ -14,6 +14,7 @@
 #include <fstream>
 #include <stdlib.h>
 #include <stdio.h>
+#include <glib.h>
 
 //for multiple processes processing
 #include <signal.h>
@@ -121,67 +122,63 @@ int mysystem(const char *cmdstr)
 
 int SAR_Profile_2D()
 {
-	//Align csi and imu data
-	if(csi_ready && imu_ready)
+	//csi_ready = false;
+	//imu_ready = false;
+	double timeDifference = fabs(t_stamp_csi-t_stamp_imu);
+	if(maxT_D < timeDifference)
 	{
-		csi_ready = false;
-		imu_ready = false;
-		double timeDifference = fabs(t_stamp_csi-t_stamp_imu);
-		if(maxT_D < timeDifference)
-		{
-			maxT_D = timeDifference;
-		}
-		//printf("T_D:%lf, ", timeDifference);
-		orientation[dataIndex % sizeLimit] = yaw;
-		CSI1[dataIndex % sizeLimit] = csi1;
-		CSI2[dataIndex % sizeLimit] = csi2;
-		if(dataIndex > 0 && dataIndex % sizeLimit == 0 && !start)
-		{
-			printf("Start!\n");
-			start = true;
-		}
-		if(start)
-		{
-			printf("max T_D:%lf, ", maxT_D);
-			start = false;
-			maxT_D = 0;
-			++count_d;
-			int r_yaw;
-			//When csi and imu data vectors reach size limit, start angle generation
-			int resolution = stepSize;      //search resolution
-      			double power = 0;
-      
-			myfile << "#" << count_d << endl;
-			
-      			for(int alpha = 0; alpha < 360; alpha += resolution)
-      			{
-				double sumpow = 0;
-        			for (int step = 0; step < stepSize; ++step)
-	      			{
-					double alpha_r = (alpha+step)*PI/180.0;
-        				double powtmp = PowerCalculation(alpha_r);
-					sumpow += powtmp;
-					////myfile << powtmp << endl;
-					//multipathProfile(alpha) += powtmp;
-					//single path
-					if(power < powtmp)
-					{
-						power= powtmp;
-						r_yaw = alpha;
-					}
-						
-      				}
-				myfile << sumpow << endl;
-			}
-			//int directPath = findDirectPath();
-			++dataIndex;
-      			//return directPath;
-			printf("Count:%d,maxPow: %0.3f, ",count_d, power);
-			return r_yaw;
-		}
-		
-		++dataIndex;
+		maxT_D = timeDifference;
 	}
+	//printf("T_D:%lf, ", timeDifference);
+	orientation[dataIndex % sizeLimit] = yaw;
+	CSI1[dataIndex % sizeLimit] = csi1;
+	CSI2[dataIndex % sizeLimit] = csi2;
+	if(dataIndex > 0 && dataIndex % sizeLimit == 0 && !start)
+	{
+		printf("Start!\n");
+		start = true;
+	}
+	if(start)
+	{
+		printf("max T_D:%lf, ", maxT_D);
+		start = false;
+		maxT_D = 0;
+		++count_d;
+		int r_yaw;
+		//When csi and imu data vectors reach size limit, start angle generation
+		int resolution = stepSize;      //search resolution
+		double power = 0;
+  
+		myfile << "#" << count_d << endl;
+			
+		for(int alpha = 0; alpha < 360; alpha += resolution)
+		{
+			double sumpow = 0;
+       			for (int step = 0; step < stepSize; ++step)
+      			{
+				double alpha_r = (alpha+step)*PI/180.0;
+       				double powtmp = PowerCalculation(alpha_r);
+				sumpow += powtmp;
+				////myfile << powtmp << endl;
+				//multipathProfile(alpha) += powtmp;
+				//single path
+				if(power < powtmp)
+				{
+					power= powtmp;
+					r_yaw = alpha;
+				}
+				
+      			}
+			myfile << sumpow << endl;
+		}
+		//int directPath = findDirectPath();
+		++dataIndex;
+      		//return directPath;
+		printf("Count:%d,maxPow: %0.3f, ",count_d, power);
+		return r_yaw;
+	}
+		
+	++dataIndex;
 
 	return -1;
 }
@@ -205,6 +202,41 @@ void csiCallback(const sar_localization::Csi::ConstPtr& msg)
   	csi_ready = true;
 }   
 // %EndTag(CALLBACK)%
+/*
+void processing()
+{
+	pid_t cpid;	
+	while(ros::ok())
+	{
+	int angle = SAR_Profile_2D();
+	if(angle > 0)
+        {
+                 printf("Alpha:%d\n", angle);
+                 //Switch to another AP
+                 switch(AP_ID)
+                 {
+                 case 0:
+                         //Switch to from AP1 to AP2
+                         AP_ID = (AP_ID+1)%AP_NUM;
+                         system("pkill -INT -n ping");   //kill the child process first
+                         system("iwconfig wlan0 essid TP5G2");
+                         printf("Switch to TP5G2 and start ping\n");
+                         cpid = mysystem("ping -q -n -i 0.01 192.168.0.3");
+                         break;
+                 case 1:
+                         AP_ID = (AP_ID+1)%AP_NUM;
+                         system("pkill -INT -n ping");      //kill the child process first
+                         system("iwconfig wlan0 essid TP5G1");
+                         printf("Switch to TP5G1 and start ping\n");
+                         cpid = mysystem("ping -q -n -i 0.01 192.168.0.2");
+                         break;
+                 }
+        }
+
+	}
+
+}
+*/
 
 int main(int argc, char **argv)
 {
@@ -212,8 +244,12 @@ int main(int argc, char **argv)
 	multipathProfile.setZero();
   	ros::NodeHandle n;
 
-  	ros::Subscriber sub1 = n.subscribe("imu", 1000, imuCallback);
-	ros::Subscriber sub2 = n.subscribe("csi", 1000, csiCallback);
+  	ros::Subscriber sub1 = n.subscribe("imu", 1, imuCallback);
+	ros::Subscriber sub2 = n.subscribe("csi", 1, csiCallback);
+
+	//ros::MultiThreadedSpinner spinner(2);
+
+	//ros::Rate(100);
 
 	//system configuration
 	//system("service network-manager stop");
@@ -247,37 +283,54 @@ int main(int argc, char **argv)
 		
 	myfile.open("power.txt");
 
-	// %Tag(SPIN)%
-        while (n.ok())
-        {
-               	ros::spinOnce();
-               	//do something
-               	int angle = SAR_Profile_2D();
-               	if(angle > 0)
-               	{
-	                printf("Alpha:%d\n", angle);
-			//Switch to another AP
-			switch(AP_ID)
+	while(ros::ok() )
+	{
+		//spinner.spinOnce();
+		ros::spinOnce();
+		if(csi_ready && imu_ready)
+		{
+			csi_ready = false;
+			imu_ready = false;
+			int angle = SAR_Profile_2D();	
+			if(angle >= 0)
 			{
-			case 0:
-				//Switch to from AP1 to AP2
-				AP_ID = (AP_ID+1)%AP_NUM;
-				system("pkill -INT -n ping");	//kill the child process first
-				system("iwconfig wlan0 essid TP5G2");
-                		printf("Switch to TP5G2 and start ping\n");
-                		cpid = mysystem("ping -q -n -i 0.05 192.168.0.3");
-				break;
-			case 1:
-				AP_ID = (AP_ID+1)%AP_NUM;
-                        	system("pkill -INT -n ping");      //kill the child process first
-                                system("iwconfig wlan0 essid TP5G1");
-                                printf("Switch to TP5G1 and start ping\n");
-	                        cpid = mysystem("ping -q -n -i 0.05 192.168.0.2");
-        	                break;	
+				printf("Alpha:%d\n", angle);
+        	                //Switch to another AP
+                	        switch(AP_ID)
+                        	{
+	                        case 0:
+        	                        //Switch to from AP1 to AP2
+                	                AP_ID = (AP_ID+1)%AP_NUM;
+                        	        system("pkill -INT -n ping");   //kill the child process first
+                                	system("iwconfig wlan0 essid TP5G2");
+                                	printf("Switch to TP5G2 and start ping\n");
+                                	cpid = mysystem("ping -q -n -i 0.05 192.168.0.3");
+                                	break;
+                        	case 1:
+                                	AP_ID = (AP_ID+1)%AP_NUM;
+                                	system("pkill -INT -n ping");      //kill the child process first
+                                	system("iwconfig wlan0 essid TP5G1");
+                                	printf("Switch to TP5G1 and start ping\n");
+                                	cpid = mysystem("ping -q -n -i 0.05 192.168.0.2");
+                                	break;
+                        	}
 			}
 		}
 	}
+
+	/*
+	GThread* process_thread;
+  	GError* err=NULL;
+
+  	if ((process_thread = g_thread_new( "data_processing", (GThreadFunc)processing, NULL)) == NULL)
+  	{
+   		printf("Failed to create serial handling thread: %s!!\n", err->message);
+    		g_error_free(err);
+  	}
+	// %Tag(SPIN)%
+        ros::spin();
 	// %EndTag(SPIN)%
+	*/
 	myfile.close();
 	system("pkill -INT -n ping");
 	return 0;
